feat: Add support for displaying file sizes in a table
This commit adds support for displaying file sizes in a table format using the `comfy-table` crate. The code now creates a `Table` object and loads the UTF8_FULL preset to set the width of the table. It then iterates over the `sizes` hashmap and adds each entry (excluding the root directory) as a row to the table. Finally, it prints out the table.