feat(cli): Add option to sort output by size
This commit adds a new command line option `--sort-by-size` that allows the user to sort the output by size. When this option is enabled, the sizes values are sorted in descending order before being printed.

The `run()` function in `cli.rs` has been modified to check if the `sort_by_size` flag is set. If it is, the sizes values are sorted using the `utils::sort_by_size()` function before being printed. Otherwise, they are printed as before.

A new utility function `sort_by_size()` has been added to `utils.rs`. This function takes a HashMap of sizes and returns a vector of key-value pairs sorted by value in descending order.

These changes improve the flexibility and usability of the CLI tool by allowing users to easily sort the output based on file or folder size.